+++ src/BTCPayServer.Lightning.All/LightningConnectionString.cs
@@ -296,25 +296,42 @@ namespace BTCPayServer.Lightning
                     }
                     break;
                 case LightningConnectionType.Eclair:
-//                    var server = Take(keyValues, "server");
-//                    if (server == null)
-//                    {
-//                        error = $"The key 'server' is mandatory for charge connection strings";
-//                        return false;
-//                    }
-//
-//                    if (server.StartsWith("//", StringComparison.OrdinalIgnoreCase))
-//                        server = "unix:" + str;
-//                    else if (server.StartsWith("/", StringComparison.OrdinalIgnoreCase))
-//                        server = "unix:/" + str;
-//
-//                    if (!Uri.TryCreate(server, UriKind.Absolute, out var uri)
-//                        || (uri.Scheme != "tcp" && uri.Scheme != "unix"))
-//                    {
-//                        error = $"The key 'server' should be an URI starting by tcp:// or unix:// or a path to the 'lightning-rpc' unix socket";
-//                        return false;
-//                    }
-//                    result.BaseUri = uri;
+                    
+                    
+                    var eclairserver = Take(keyValues, "server");
+                    
+                    if (eclairserver == null)
+                    {
+                        error = $"The key 'server' is mandatory for lnd connection strings";
+                        return false;
+                    }
+                    if (!Uri.TryCreate(eclairserver, UriKind.Absolute, out var eclairuri)
+                        || (eclairuri.Scheme != "http" && eclairuri.Scheme != "https"))
+                    {
+                        error = $"The key 'server' should be an URI starting by http:// or https://";
+                        return false;
+                    }
+
+                    result.BaseUri = eclairuri;
+                    result.Password = Take(keyValues, "password");
+                    result.BitcoinHost  = Take(keyValues, "bitcoin-host");
+
+                    if (result.BitcoinHost == null)
+                    {
+                        error = $"The key 'bitcoin-host' is mandatory for eclair connection strings";
+                        return false;
+                    }
+                    
+                    
+
+                    result.BitcoinAuth= Take(keyValues, "bitcoin-auth");
+                    
+                    if (result.BitcoinAuth == null)
+                    {
+                        error = $"The key 'bitcoin-auth' is mandatory for eclair connection strings";
+                        return false;
+                    }
+                    
                     break;
                 default:
                     throw new NotSupportedException(connectionType.ToString());
@@ -424,6 +441,9 @@ namespace BTCPayServer.Lightning
         public string CookieFilePath { get; set; }
         public string MacaroonDirectoryPath { get; set; }
 
+        public string BitcoinHost { get; set; }
+        public string BitcoinAuth { get; set; }
+        
         public Uri ToUri(bool withCredentials)
         {
             if (withCredentials)
@@ -495,6 +515,26 @@ namespace BTCPayServer.Lightning
                         builder.Append($";allowinsecure=true");
                     }
                     break;
+                case LightningConnectionType.Eclair:
+                    builder.Append($";server={BaseUri}");
+                    if (Password != null)
+                    {
+                        builder.Append($";password={Password}");
+                    }
+                    if (Password != null)
+                    {
+                        builder.Append($";password={Password}");
+                    }
+                    if (BitcoinHost != null)
+                    {
+                        builder.Append($";bitcoin-host={BitcoinHost}");
+                    }
+                    if (BitcoinAuth != null)
+                    {
+                        builder.Append($";bitcoin-auth={BitcoinAuth}");
+                    }
+                    
+                    break; 
                 default:
                     throw new NotSupportedException(type);
             }
