+++ src/BTCPayServer.Lightning.Eclair/EclairLightningClient.cs
@@ -1,237 +1,24 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Globalization;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Text;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using BTCPayServer.Lightning.Eclair.Models;
 using NBitcoin;
 using NBitcoin.RPC;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-using Newtonsoft.Json.Serialization;
 using WebSocketSharp;
 
 namespace BTCPayServer.Lightning.Eclair
 {
-    public class EclairClient()
-    {
-        private Uri _Address;
-        private HttpClient _httpClient;
-
-
-        public EclairClient(Uri address, string password)
-        {
-            if (address == null)
-                throw new ArgumentNullException(nameof(address));
-
-            _Address = address;
-
-            _httpClient = new HttpClient();
-            _httpClient.BaseAddress = address;
-            _httpClient.DefaultRequestHeaders.Accept.Clear();
-            _httpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Basic",
-                Convert.ToBase64String(Encoding.Default.GetBytes($":{password}")));
-        }
-
-        public async Task<GetInfoResponse> GetInfo(CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<GetInfoResponse>("getinfo", cts);
-        }
-
-        public async Task<string> ConnectToNode(string nodeId, string host, int port,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("connect", cts, nodeId, host, port);
-        }
-
-        public async Task<string> ConnectToNode(string uri, CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("connect", cts, uri);
-        }
-
-        public async Task<string> OpenChannel(string nodeId, long fundingSatoshis, long pushMsat = 0,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("open", cts, nodeId, fundingSatoshis, pushMsat);
-        }
-
-        public async Task<string> OpenChannel(string nodeId, long fundingSatoshis, long pushMsat,
-            long feerateSatPerByte, CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("open", cts, nodeId, fundingSatoshis, pushMsat, feerateSatPerByte);
-        }
-
-        public async Task<string> UpdateRelayFee(string channelId, long feeBaseMsat, long feeProportionalMillionths,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("updaterelayfee", cts, channelId, feeBaseMsat,
-                feeProportionalMillionths);
-        }
-
-        public async Task<ListChannelsResponseItem[]> ListChannels(CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<ListChannelsResponseItem[]>("channels", cts);
-        }
-
-        public async Task<ListChannelsResponseItem[]> ListChannels(string nodeId,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<ListChannelsResponseItem[]>("channels", cts, nodeId);
-        }
-
-        public async Task<ListChannelsResponseItem> GetChannel(string channelId,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<ListChannelsResponseItem>("channel", cts, channelId);
-        }
-
-        public async Task<ListChannelsResponseItem> ListAllChannels(CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<ListChannelsResponseItem>("allchannels", cts);
-        }
-
-        public async Task<NodeResult[]> ListAllNodes(CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<NodeResult[]>("allnodes", cts);
-        }
-
-        public async Task<string> Receive(string description, CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("receive", cts, description);
-        }
-
-        public async Task<string> Receive(string description, long mSat,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("receive", cts, mSat, description);
-        }
-
-        public async Task<string> Receive(string description, long mSat, int expirySeconds,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("receive", cts,
-                mSat,
-                description,
-                expirySeconds);
-        }
-
-        public async Task<SendResponse> Send(string paymentRequest, CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<SendResponse>("send", cts, paymentRequest);
-        }
-
-        public async Task<SendResponse> Send(string paymentRequest, long amountMsat,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<SendResponse>("send", cts, paymentRequest, amountMsat);
-        }
-
-        public async Task<SendResponse> Send(long amountMsat, string paymentHash, string nodeId,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<SendResponse>("send", cts, amountMsat, paymentHash, nodeId);
-        }
-
-        public async Task<CheckInvoiceResponse> CheckInvoice(string paymentRequest,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<CheckInvoiceResponse>("checkinvoice", cts, paymentRequest);
-        }
-
-        public async Task<CheckInvoiceResponse> ParseInvoice(string paymentRequest,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<CheckInvoiceResponse>("parseinvoice", cts, paymentRequest);
-        }
-
-        public async Task<bool> CheckPayment(string paymentRequestOrHash,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<bool>("checkpayment", cts, paymentRequestOrHash);
-        }
-
-        public async Task<string> Close(string channelId, CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("close", cts, channelId);
-        }
-
-        public async Task<string> Close(string channelId, string scriptPubKey,
-            CancellationToken cts = default(CancellationToken))
-        {
-            return await SendCommandAsync<string>("close", cts, channelId, scriptPubKey);
-        }
-
-        private async Task<T> SendCommandAsync<T>(string method, CancellationToken cts, params object[] parameters)
-        {
-            var jsonSerializer = new JsonSerializerSettings
-                {ContractResolver = new CamelCasePropertyNamesContractResolver()};
-
-            var body = new JsonRpcCommand(method, parameters);
-
-            var request = new HttpRequestMessage(HttpMethod.Post, string.Empty);
-            request.Content = new StringContent(body.ToString(jsonSerializer), Encoding.UTF8, "application/json");
-
-            var rawResult = await _httpClient.SendAsync(request, cts);
-            var rawJson = await rawResult.Content.ReadAsStringAsync();
-            var result = JObject.Parse(rawJson).ToObject<JsonRpcResult<T>>(JsonSerializer.Create(jsonSerializer));
-            if (result.Error != null && !string.IsNullOrEmpty(result.Error.Message))
-            {
-                throw new InvalidOperationException(result.Error.Message);
-            }
-
-            return result.Result;
-        }
-
-
-        internal class JsonRpcResult<T>
-        {
-            public class JsonRpcResultError
-            {
-                [JsonProperty("code")] public int Code { get; set; }
-                [JsonProperty("message")] public string Message { get; set; }
-            }
-
-            [JsonProperty("result")] public T Result { get; set; }
-            [JsonProperty("error")] public JsonRpcResultError Error { get; set; }
-            [JsonProperty("id")] public string Id { get; set; }
-        }
-
-        internal class JsonRpcCommand
-        {
-            [JsonProperty("jsonRpc")] public string JsonRpc { get; set; } = "2.0";
-            [JsonProperty("id")] public string Id { get; set; } = Guid.NewGuid().ToString();
-            [JsonProperty("method")] public string Method { get; set; }
-
-            [JsonProperty("params")] public object[] Parameters { get; set; }
-
-            public JsonRpcCommand()
-            {
-            }
-
-            public JsonRpcCommand(string method, object[] parameters)
-            {
-                Method = method;
-                Parameters = parameters;
-            }
-
-            public string ToString(JsonSerializerSettings jsonSerializer)
-            {
-                return JObject.FromObject(this, JsonSerializer.Create(jsonSerializer)).ToString();
-            }
-        }
-    }
-
-
     public class EclairLightningClient : ILightningClient
     {
         private readonly Uri _address;
         private readonly Network _network;
         private readonly RPCClient _rpcClient;
-        private readonly ConcurrentDictionary<string, LightningInvoice> _memoryInvoices = new ConcurrentDictionary<string, LightningInvoice>();
+
+        private readonly ConcurrentDictionary<string, LightningInvoice> _memoryInvoices =
+            new ConcurrentDictionary<string, LightningInvoice>();
+
         private EclairClient _eclairClient;
 
         public EclairLightningClient(Uri address, string password, Network network, RPCClient rpcClient)
@@ -250,7 +37,7 @@ namespace BTCPayServer.Lightning.Eclair
         public async Task<LightningInvoice> GetInvoice(string invoiceId,
             CancellationToken cancellation = default(CancellationToken))
         {
-         return await  UpdateLocalLightningInvoice(invoiceId, cancellation);
+            return await UpdateLocalLightningInvoice(invoiceId, cancellation);
         }
 
         public async Task<LightningInvoice> CreateInvoice(LightMoney amount, string description, TimeSpan expiry,
@@ -274,14 +61,11 @@ namespace BTCPayServer.Lightning.Eclair
 
             _memoryInvoices.TryAdd(invoice.Id, invoice);
             return invoice;
-
         }
 
         public async Task<ILightningInvoiceListener> Listen(CancellationToken cancellation = default(CancellationToken))
         {
-
-            var listener = new EclairWebsocketListener(this);
-            throw new Exception();
+            return new EclairWebsocketListener(this);
         }
 
         public async Task<LightningNodeInformation> GetInfo(CancellationToken cancellation = default(CancellationToken))
@@ -301,6 +85,12 @@ namespace BTCPayServer.Lightning.Eclair
             try
             {
                 var sendResult = await _eclairClient.Send(bolt11, cancellation);
+                if (sendResult.failures != null && sendResult.failures.Any(item =>
+                        item.t.Equals("route not found", StringComparison.InvariantCultureIgnoreCase)))
+                {
+                    return new PayResponse(PayResult.CouldNotFindRoute);
+                }
+
                 return new PayResponse(PayResult.Ok);
             }
             catch (Exception)
@@ -312,12 +102,28 @@ namespace BTCPayServer.Lightning.Eclair
         public async Task<OpenChannelResponse> OpenChannel(OpenChannelRequest openChannelRequest,
             CancellationToken cancellation = default(CancellationToken))
         {
-            var result = await _eclairClient.OpenChannel(openChannelRequest.NodeInfo.NodeId.ToString(),
-                openChannelRequest.ChannelAmount.Satoshi
-                , 0,
-                Convert.ToInt64(openChannelRequest.FeeRate.SatoshiPerByte), cancellation);
+            try
+            {
+                var result = await _eclairClient.OpenChannel(openChannelRequest.NodeInfo.NodeId.ToString(),
+                    openChannelRequest.ChannelAmount.Satoshi
+                    , 0,
+                    Convert.ToInt64(openChannelRequest.FeeRate.SatoshiPerByte), cancellation);
 
-            return new OpenChannelResponse(OpenChannelResult.Ok);
+                if (result.Contains("already exists"))
+                {
+                    return new OpenChannelResponse(OpenChannelResult.AlreadyExists); 
+                }
+                
+                return new OpenChannelResponse(OpenChannelResult.Ok);
+            }
+            catch (Exception e)
+            {
+                if (e.Message.Contains("command failed: not connected"))
+                {
+                    return new OpenChannelResponse(OpenChannelResult.PeerNotConnected);
+                }
+                return  new OpenChannelResponse(OpenChannelResult.AlreadyExists);
+            }
         }
 
         public async Task<BitcoinAddress> GetDepositAddress()
@@ -329,16 +135,18 @@ namespace BTCPayServer.Lightning.Eclair
         {
             await _eclairClient.ConnectToNode(nodeInfo.NodeId.ToString(), nodeInfo.Host, nodeInfo.Port);
         }
-        
-        private async Task<LightningInvoice> UpdateLocalLightningInvoice(string paymentHash,CancellationToken cancellation)
+
+        private async Task<LightningInvoice> UpdateLocalLightningInvoice(string paymentHash,
+            CancellationToken cancellation)
         {
             var paid = await _eclairClient.CheckPayment(paymentHash, cancellation);
-            if (!_memoryInvoices.ContainsKey(paymentHash)) return new LightningInvoice()
-            {
-                Id = paymentHash,
-                Status = paid? LightningInvoiceStatus.Paid: LightningInvoiceStatus.Unpaid,
-                PaidAt = paid? DateTimeOffset.Now : (DateTimeOffset?) null
-            };
+            if (!_memoryInvoices.ContainsKey(paymentHash))
+                return new LightningInvoice()
+                {
+                    Id = paymentHash,
+                    Status = paid ? LightningInvoiceStatus.Paid : LightningInvoiceStatus.Unpaid,
+                    PaidAt = paid ? DateTimeOffset.Now : (DateTimeOffset?) null
+                };
             var invoice = _memoryInvoices[paymentHash];
             if (invoice.ExpiresAt <= DateTimeOffset.Now && !paid)
             {
@@ -355,7 +163,6 @@ namespace BTCPayServer.Lightning.Eclair
             }
 
             return invoice;
-
         }
 
         public class EclairWebsocketListener : ILightningInvoiceListener
@@ -379,8 +186,9 @@ namespace BTCPayServer.Lightning.Eclair
             private void Connect()
             {
                 _websocketConnection = new WebSocket(_address);
-                _websocketConnection.Connect();
                 _websocketConnection.OnMessage += WebsocketConnectionOnOnMessage;
+                _websocketConnection.OnError += (sender, args) => { Console.WriteLine(args.Message); };
+                _websocketConnection.Connect();
             }
 
             private void WebsocketConnectionOnOnMessage(object sender, MessageEventArgs e)
@@ -400,6 +208,12 @@ namespace BTCPayServer.Lightning.Eclair
             {
                 while (!cancellation.IsCancellationRequested)
                 {
+                    if (!_websocketConnection.IsAlive)
+                    {
+                        Dispose();
+                        Connect();
+                    }
+
                     if (_receivedInvoiceQueue.IsEmpty)
                     {
                         await Task.Delay(TimeSpan.FromMilliseconds(500), cancellation);
@@ -409,13 +223,10 @@ namespace BTCPayServer.Lightning.Eclair
                     if (!_receivedInvoiceQueue.TryDequeue(out var paymentHash)) continue;
 
                     return await _eclairLightningClient.UpdateLocalLightningInvoice(paymentHash, cancellation);
-                    
                 }
 
                 return null;
             }
-
-            
         }
     }
 }
