+++ tests/Tester.cs
@@ -51,11 +51,11 @@ namespace BTCPayServer.Lightning.Tests
 
         public static EclairLightningClient CreateEclairClient()
         {
-            return new EclairLightningClient(new Uri("http://127.0.0.1:4570"), "bukkake", Network);
+            return new EclairLightningClient(new Uri("http://127.0.0.1:4570"), "bukkake", Network, CreateRPC());
         }  
         public static EclairLightningClient CreateEclairClientDest()
         {
-            return new EclairLightningClient(new Uri("http://127.0.0.1:4571"), "bukkake", Network);
+            return new EclairLightningClient(new Uri("http://127.0.0.1:4571"), "bukkake", Network, CreateRPC());
         }
         
 
@@ -70,24 +70,24 @@ namespace BTCPayServer.Lightning.Tests
 
         public static IEnumerable<ILightningClient> GetLightningClients()
         {
-            yield return CreateChargeClient();
-            yield return CreateCLightningClient();
-            yield return CreateLndClient();
+//            yield return CreateChargeClient();
+//            yield return CreateCLightningClient();
+//            yield return CreateLndClient();
             yield return CreateEclairClient();
         }
 
         public static IEnumerable<ILightningClient> GetLightningSenderClients()
         {
-            yield return CreateCLightningClient();
-            yield return CreateLndClient();
+//            yield return CreateCLightningClient();
+//            yield return CreateLndClient();
             
             yield return CreateEclairClient();
             
         }
         public static IEnumerable<ILightningClient> GetLightningDestClients()
         {
-            yield return CreateCLightningClientDest();
-            yield return CreateLndClientDest();
+//            yield return CreateCLightningClientDest();
+//            yield return CreateLndClientDest();
             yield return CreateEclairClientDest();
         }
     }
