+++ tests/docker-compose.yml
@@ -161,6 +161,7 @@ services:
         -Declair.headless
         -Declair.chain=regtest
         -Declair.server.binding-ip=0.0.0.0
+        -Declair.server.public-ips.0=10.5.0.1
         -Declair.server.port=9735
         -Declair.api.enabled=true
         -Declair.api.binding-ip=0.0.0.0
@@ -183,6 +184,10 @@ services:
       - "eclair_dir:/data"
     links:
       - bitcoind
+    networks:
+      - default:
+      - eclairhax:
+          - ipv4_address: 10.5.0.1
 
   eclair_dest:
     restart: unless-stopped
@@ -194,6 +199,7 @@ services:
         -Declair.headless
         -Declair.chain=regtest
         -Declair.server.binding-ip=0.0.0.0
+        -Declair.server.public-ips.0=10.5.0.2
         -Declair.server.port=9735
         -Declair.api.enabled=true
         -Declair.api.binding-ip=0.0.0.0
@@ -216,6 +222,11 @@ services:
       - "eclair_dest_dir:/data"
     links:
       - bitcoind
+    networks:
+      - default:
+      - eclairhax:
+          - ipv4_address: 10.5.0.2
+      
 
 volumes:
     lnd_dir:
@@ -227,3 +238,10 @@ volumes:
     eclair_dir:
     eclair_dest_dir:
 
+networks:
+  default:
+  eclairhax:
+    driver: bridge
+    ipam:
+      config:
+        - subnet: 10.5.0.0/16
